requirements.txt
pyTelegramBotAPI~=4.10.0
openpyxl~=3.1.1
python-dateutil~=2.8.2